requirements/base.txt
beautifulsoup4==4.7.1
celery==4.2.1
dj-database-url==0.5.0
Django==2.1.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-ratelimit==2.0.0
geoip2==2.9.0
html2text==2018.1.9
psycopg2-binary==2.8.6
python-decouple==3.1
pytz==2018.9
requests==2.21.0
